assignment2/requirements.txt
numpy==1.13.3
seaborn==0.8.1
matplotlib==2.0.2
networkx==1.11
python_igraph==0.7.1.post6
